Extract unit tests for even hole free
