+++ inducer/EvenHoleFree.py
@@ -7,7 +7,6 @@ from networkx.algorithms import graph_clique_number
 from networkx import complement
 from inducer.helper import make_cycle
 from inducer.container import induced_subgraph
-import unittest
 
 
 def even_hole_free(G):
@@ -34,39 +33,3 @@ def odd_hole_free(G):
             free = induced
         i += 2
     return free
-
-
-class Test(unittest.TestCase):
-
-    def setUp(self):
-        pass
-
-    def tearDown(self):
-        pass
-
-    def testEvenHoleFree(self):
-        g = make_cycle(4)
-        self.assertEqual(even_hole_free(g).edges(), g.edges())
-        self.assertEqual(even_hole_free(g).nodes(), g.nodes())
-        g = make_cycle(5)
-        self.assertEqual(even_hole_free(g), None)
-        g = make_cycle(10)
-        self.assertEqual(even_hole_free(g).edges(), g.edges())
-        self.assertEqual(even_hole_free(g).nodes(), g.nodes())
-
-    def testOddHoleFree(self):
-        g = make_cycle(4)
-        self.assertEqual(odd_hole_free(g), None)
-        g = make_cycle(5)
-        self.assertEqual(odd_hole_free(g).edges(), g.edges())
-        self.assertEqual(odd_hole_free(g).nodes(), g.nodes())
-        g = make_cycle(10)
-        self.assertEqual(odd_hole_free(g), None)
-        g = make_cycle(11)
-        self.assertEqual(odd_hole_free(g).edges(), g.edges())
-        self.assertEqual(odd_hole_free(g).nodes(), g.nodes())
-
-
-if __name__ == "__main__":
-    # import sys;sys.argv = ['', 'Test.testName']
-    unittest.main()
